--- src/main/java/com/carero/dto/RecruitPostResponseDto.java
@@ -1,8 +1,0 @@
-package com.carero.dto;
-
-import lombok.Data;
-
-@Data
-public class RecruitPostResponseDto {
-    private Long id;
-}
