+++ src/main/java/com/carero/dto/recruit/CreateRecruitResponseDto.java
@@ -1,0 +1,15 @@
+package com.carero.dto.recruit;
+
+import lombok.Data;
+
+@Data
+public class CreateRecruitResponseDto {
+    public CreateRecruitResponseDto() {
+    }
+
+    private Long id;
+
+    public CreateRecruitResponseDto(Long id) {
+        this.id = id;
+    }
+}
